Fixer le conflit de fichier + fait fonctionner tout
