+++ src/main/java/ikasaidi/backend_lab/BackendLabApplication.java
@@ -3,10 +3,7 @@ package ikasaidi.backend_lab;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-@SpringBootApplication(exclude = {
-        org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration.class,
-        org.springframework.boot.autoconfigure.orm.jpa.HibernateJpaAutoConfiguration.class
-})
+@SpringBootApplication
 public class BackendLabApplication {
 
     public static void main(String[] args) {
