+++ src/main/java/ikasaidi/backend_lab/repositories/PersonRepository.java
@@ -1,0 +1,12 @@
+package ikasaidi.backend_lab.repositories;
+
+import ikasaidi.backend_lab.models.Person;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+
+@Repository
+public interface PersonRepository extends JpaRepository<Person, Integer> {
+    List<Person> findByNameContainingIgnoreCase(String name);
+}
