--- src/main/java/rachelsil/lab01/Lab01Application.java
@@ -1,13 +1,0 @@
-package rachelsil.lab01;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Lab01Application {
-
-	public static void main(String[] args) {
-		SpringApplication.run(Lab01Application.class, args);
-	}
-
-}
