+++ src/main/resources/application.properties
@@ -1,16 +1,10 @@
 spring.application.name=Backend_lab
 server.port=8585
 
-#spring.datasource.driver-class-name=org.mariadb.jdbc.Driver
-#spring.datasource.url=jdbc:mariadb://localhost:3306/lab1
-#spring.datasource.username=root
-#spring.datasource.password=
-#
-#spring.jpa.hibernate.ddl-auto=create-drop
-#spring.jpa.show-sql=true
-#spring.jackson.date-format=yyyy-MM-dd
-#spring.jackson.serialization.write-dates-as-timestamps:false
-#spring.jackson.time-zone:EST
-#
-#spring.jpa.defer-datasource-initialization=true
-#spring.sql.init.mode=always
+spring.datasource.url=jdbc:sqlite:persons.db
+spring.datasource.driver-class-name=org.sqlite.JDBC
+
+spring.jpa.database-platform=org.hibernate.community.dialect.SQLiteDialect
+spring.jpa.hibernate.ddl-auto=update
+spring.datasource.username=
+spring.datasource.password=
